// src/types/geocoding.ts
interface GeocodingResult {
  id: number;
  name: string;
  latitude: number;
  longitude: number;
  elevation: number;
  feature_code: string;
  country_code: string;
  admin1_id: number;
  admin2_id: number;
  admin3_id: number;
  admin4_id: number;
  timezone: string;
  population: number;
  postcodes: string[];
  country_id: number;
  country: string;
  admin1: string;
  admin2: string;
  admin3: string;
  admin4: string;
}

interface GeocodingSuccessResponse {
  results: GeocodingResult[];
  error?: false;
  reason?: never;
}

interface GeocodingErrorResponse {
  error: true;
  reason: string;
  results?: never;
}

type GeocodingResponse = GeocodingSuccessResponse | GeocodingErrorResponse;

export type {
  GeocodingResult,
  GeocodingSuccessResponse,
  GeocodingErrorResponse,
  GeocodingResponse,
};